+++ TestLibTracer/Scene/TestPixel.cs
@@ -10,9 +10,9 @@ namespace TestLibTracer.Scene
         private const Int32 SOME_X = 123;
         private const Int32 SOME_Y = 321;
 
-        private const Single SOME_RED = 0.1f;
-        private const Single SOME_GREEN = 0.2f;
-        private const Single SOME_BLUE = 0.3f;
+        private const Double SOME_RED = 0.1;
+        private const Double SOME_GREEN = 0.2;
+        private const Double SOME_BLUE = 0.3;
 
         private Pixel _pixel;
